\name{sequence.table}
\alias{sequence.table}
\docType{data}
\title{
Counts of amplicon sequence variants in 8 sponge samples of Svenzai zeai. Half were adults, and half were larva.
}
\description{
8 samples of the sponge Svenzai seai were collected for 16 rRNA sequencing of their microbiomes. This table is one of the outputs of the dada2 pipeline which works with next-genration sequencing data. The rows are the sample, and the columns contain counts of some amplicon sequence variant found in some sample. 
}
\usage{data(sequence.table)}
\format{
A 8 X 2041 table of sequence count information}
\source{
Easson Lab, Middle Tennessee State University
}


