\name{seqwork-package}
\alias{seqwork-package}
\alias{seqwork}
\docType{package}
\title{
\packageTitle{seqwork}
}
\description{
\packageDescription{seqwork}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{seqwork}
\packageIndices{seqwork}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{seqwork}

Maintainer: \packageMaintainer{seqwork}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
