\name{taxa}
\alias{taxa}
\docType{data}
\title{
Taxanomic idenitity for 2041 amplicon sequence variants
}
\description{
This data is from an Illumina MiSeq run conducted at Middle Tenessee State University in March 2022. These amplicon sequence variants were identified from 35 samples using the dada2 pipeline. 
}
\usage{data(taxa)}
\format{
A 2041 X 7 table of nucelotide sequences as rows, and the taxanomic identity assigned to that sequence in the columns.}
\source{
Easson Lab, Middle Tennessee State University
}


